+++ examples/msgpack.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 	"os"
-	"strings"
 
 	// "strings"
 
@@ -13,7 +12,7 @@ import (
 func main() {
 
 	folderPaths := []string{
-		utils.ConvertPathToOSFormat("C:\\Users\\Restop-1294\\My_Apps\\chrome-extensions\\modify_chatgpt_prompts\\extension"),
+		utils.ConvertPathToOSFormat("C:\\Users\\Restop-1294\\My_Apps\\Windmillcode_app_tutorials\\tutorials\\working_with_peerjs_for_flutter\\msgpack-javascript-2.8.0\\src"),
 	}
 
 	files := []string{}
@@ -26,7 +25,7 @@ func main() {
 				files = append(files, path)
 			},
 			Filter: func(path string, info os.FileInfo) bool {
-				return  !strings.HasSuffix(path, ".png") && !strings.Contains(path,".git") && !strings.Contains(path,".md");
+				return true
 			},
 		}