+++ examples/video_sdk_flutter_webrtc_app.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 	"os"
-	"strings"
 
 	// "strings"
 
@@ -13,7 +12,9 @@ import (
 func main() {
 
 	folderPaths := []string{
-		utils.ConvertPathToOSFormat("C:\\Users\\Restop-1294\\My_Apps\\chrome-extensions\\modify_chatgpt_prompts\\extension"),
+		utils.ConvertPathToOSFormat(
+			"C:\\Users\\Restop-1294\\My_Apps\\Windmillcode_app_tutorials\\tutorials\\working_with_peerjs_for_flutter\\webrtc\\modern_webrtc\\lib",
+		),
 	}
 
 	files := []string{}
@@ -22,11 +23,11 @@ func main() {
 			RootDir: folderPath, // Specify your directory here
 			Predicate: func(path string, info os.FileInfo) {
 				// Action to perform on each .dart file that is not a _test.dart or g.dart file
-				fmt.Println("Found  file:", path)
+				fmt.Println("Found dart file:", path)
 				files = append(files, path)
 			},
 			Filter: func(path string, info os.FileInfo) bool {
-				return  !strings.HasSuffix(path, ".png") && !strings.Contains(path,".git") && !strings.Contains(path,".md");
+				return true
 			},
 		}